Move Expression to new package me.willkroboth.ConfigCommands.RegisteredCommands.Expressions
Move subclasses of Expression out of Expression.java
